+++ CHANGES.rst
@@ -3,8 +3,8 @@ Changelog
 
 1.3.1 (unreleased)
 ----------------
-- Fixed ddetail template
-- Fixed a document generation bug when result indicator is None
+- Fix ddetail template
+- Fix a document generation bug when result indicator is None
   [fngaha]
 - Enabled categories column on action dashboard
   [fngaha]